Tests: Merge test_collectioneditor.py and test_collectionseditor.py
